+++ src/main/webui/src/ProposalEditorView/submitProposal/submitPanel.tsx
